# api/docs/swagger.yaml
definitions:
...
  auth_service.AddUserRelationRequest:
    properties:
      relation_id:
        type: string
      user_id:
        type: string
    type: object
...
  auth_service.CreateUserRequest:
    properties:
      active:
        type: integer
      client_platform_id:
        type: string
      client_type_id:
        type: string
      email:
        type: string
      expires_at:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      photo_url:
        type: string
      project_id:
        type: string
      role_id:
        type: string
    type: object
...
  auth_service.GetUserListResponse:
    properties:
      count:
        type: integer
      users:
        items:
          $ref: '#/definitions/auth_service.User'
        type: array
    type: object
...
  auth_service.ResetPasswordRequest:
    properties:
      password:
        type: string
      token:
        type: string
      user_id:
        type: string
    type: object
  auth_service.Role:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
    type: object
...
  auth_service.SendMessageToEmailRequest:
    properties:
      base_url:
        type: string
      email:
        type: string
      token:
        type: string
    type: object
...
  auth_service.UpdateUserRequest:
    properties:
      active:
        type: integer
      client_platform_id:
        type: string
      client_type_id:
        type: string
      email:
        type: string
      expires_at:
        type: string
      id:
        type: string
      login:
        type: string
      name:
        type: string
      phone:
        type: string
      photo_url:
        type: string
      project_id:
        type: string
      role_id:
        type: string
    type: object
...
  auth_service.User:
    properties:
      active:
        type: integer
      client_platform_id:
        type: string
      client_type_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      expires_at:
        type: string
      id:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      photo_url:
        type: string
      project_id:
        type: string
      role_id:
        type: string
      updated_at:
        type: string
    type: object
...
  auth_service.UserRelation:
    properties:
      relation_id:
        type: string
      user_id:
        type: string
    type: object
  auth_service.UserRelationPrimaryKey:
    properties:
      relation_id:
        type: string
      user_id:
        type: string
    type: object
...
  http.Response:
    properties:
      data:
        type: object
      description:
        type: string
      status:
        type: string
    type: object
...
  models.StructBody:
    properties:
      body:
        additionalProperties: true
        type: object
    type: object
...
info:
  contact: {}
  description: This is a api gateway
  license: {}
  termsOfService: https://udevs.io
paths:
...
  /upsert-user-info/{user-id}:
    post:
      consumes:
      - application/json
      description: Upsert UserInfo
      operationId: upsert_user_info
      parameters:
      - description: UpsertUserInfoRequestBody
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.StructBody'
      - description: user-id
        in: path
        name: user-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Role data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.Role'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Upsert UserInfo
      tags:
      - UpsertUserInfo
  /user:
    get:
      consumes:
      - application/json
      description: Get User List
      operationId: get_user_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - description: client-platform-id
        in: query
        name: client-platform-id
        type: string
      - description: client-type-id
        in: query
        name: client-type-id
        type: string
      - description: project-id
        in: query
        name: project-id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetUserListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.GetUserListResponse'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get User List
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create User
      operationId: create_user
      parameters:
      - description: CreateUserRequestBody
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth_service.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create User
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update User
      operationId: update_user
      parameters:
      - description: UpdateUserRequestBody
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth_service.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update User
      tags:
      - User
...
  /user-relation:
    delete:
      consumes:
      - application/json
      description: Get UserRelation
      operationId: delete_user_relation
      parameters:
      - description: UserRelationPrimaryKeyBody
        in: body
        name: user-relation
        required: true
        schema:
          $ref: '#/definitions/auth_service.UserRelationPrimaryKey'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete UserRelation
      tags:
      - UserRelation
    post:
      consumes:
      - application/json
      description: Create UserRelation
      operationId: add_user_relation
      parameters:
      - description: AddUserRelationRequestBody
        in: body
        name: user-relation
        required: true
        schema:
          $ref: '#/definitions/auth_service.AddUserRelationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: UserRelation data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.UserRelation'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create UserRelation
      tags:
      - UserRelation
  /user/{user-id}:
    delete:
      consumes:
      - application/json
      description: Get User
      operationId: delete_user
      parameters:
      - description: user-id
        in: path
        name: user-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete User
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Get User By ID
      operationId: get_user_by_id
      parameters:
      - description: user-id
        in: path
        name: user-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: UserBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get User By ID
      tags:
      - User
  /user/reset-password:
    put:
      consumes:
      - application/json
      description: Reset Password
      operationId: reset_password
      parameters:
      - description: ResetPasswordRequestBody
        in: body
        name: reset_password
        required: true
        schema:
          $ref: '#/definitions/auth_service.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update User
      tags:
      - User
  /user/send-message:
    post:
      consumes:
      - application/json
      description: Send Message to User Email
      operationId: send_message_to_user_email
      parameters:
      - description: SendMessageToEmailRequestBody
        in: body
        name: send_message
        required: true
        schema:
          $ref: '#/definitions/auth_service.SendMessageToEmailRequest'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Send Message To User
      tags:
      - User
...
  /v2/user:
    get:
      consumes:
      - application/json
      description: Get User List
      operationId: get_user_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - description: client-platform-id
        in: query
        name: client-platform-id
        type: string
      - description: client-type-id
        in: query
        name: client-type-id
        type: string
      - description: project-id
        in: query
        name: project-id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetUserListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.GetUserListResponse'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get User List
      tags:
      - V2_User
    post:
      consumes:
      - application/json
      description: Create User
      operationId: create_user
      parameters:
      - description: CreateUserRequestBody
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth_service.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create User
      tags:
      - V2_User
    put:
      consumes:
      - application/json
      description: Update User
      operationId: update_user
      parameters:
      - description: UpdateUserRequestBody
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth_service.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update User
      tags:
      - V2_User
...
  /v2/user/{user-id}:
    delete:
      consumes:
      - application/json
      description: Get User
      operationId: delete_user
      parameters:
      - description: user-id
        in: path
        name: user-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete User
      tags:
      - V2_User
    get:
      consumes:
      - application/json
      description: Get User By ID
      operationId: get_user_by_id
      parameters:
      - description: user-id
        in: path
        name: user-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: UserBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get User By ID
      tags:
      - V2_User
...
swagger: "2.0"